Move Report Extension on GitHub to Functions View
