+++ package.json
@@ -141,7 +141,7 @@
     "onCommand:service.explorer.openFile",
     "onCommand:function.openInEditor",
     "onCommand:service.explorer.refresh",
-    "onCommand:service.explorer.reportIssue",
+    "onCommand:function.explorer.reportIssue",
     "onCommand:service.explorer.apply",
     "onCommand:service.explorer.deleteLocal",
     "onCommand:service.output",
@@ -294,7 +294,7 @@
         "icon": "$(refresh)"
       },
       {
-        "command": "service.explorer.reportIssue",
+        "command": "function.explorer.reportIssue",
         "title": "Report Extension Issue on GitHub",
         "category": "Knative",
         "icon": "$(issues)"
@@ -609,6 +609,11 @@
       ],
       "view/title": [
         {
+          "command": "function.explorer.reportIssue",
+          "when": "view == knativeFunctionProjectExplorer",
+          "group": "navigation@4"
+        },
+        {
           "command": "function.explorer.create",
           "when": "view == knativeFunctionProjectExplorer",
           "group": "navigation@0"
@@ -642,11 +647,6 @@
           "command": "service.explorer.refresh",
           "when": "view == knativeServingProjectExplorer",
           "group": "navigation@1"
-        },
-        {
-          "command": "service.explorer.reportIssue",
-          "when": "view == knativeServingProjectExplorer",
-          "group": "navigation@2"
         },
         {
           "command": "knative.version",
